Mutations and copy number from colorectal adenocarcinomas (PMID: 30177804). n = 19 patients, multi-region whole-exome sequencing. Calls generated from Cross et al. (PMID: 30177804).

Mutations and copy number from colorectal adenocarcinomas (PMID: 30177804). n = 19 patients, multi-region whole-exome sequencing. Calls generated from Cross et al. (PMID: 30177804).

data(CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018)

Format

Data for the adenocarcinoma cohort by Cross et al. The data has been provided by WCh Cross, as of the Nature Ecology and Evolution paper (PMID: 30177804). Mutations are reported as present or absent in each one of the sequenced biopsies. All the data (calls, definition of driver events, etc.) is as provided by the authos, and as used in the paper Caravagna et al (PMID: 30171232) to show repeated evolution in breast cancers with REVOLVER.

References

Cross, W., Kovac, M., Mustonen, V., Temko, D., Davis, H., Baker, A.M., Biswas, S., Arnold, R., Chegwidden, L., Gatenbee, C. et al., 2018. The evolutionary landscape of colorectal tumorigenesis. Nature Ecology and Evolution, 2(10), p.1661. PMID 30177804.

Examples

data(CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018)
CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018

#> # A tibble: 63 x 7
#>    patientID variantID CCF                    is.clonal is.driver Misc   cluster
#>    <chr>     <chr>     <chr>                  <lgl>     <lgl>     <chr>  <chr>  
#>  1 adenoma_1 ACVR2A    R1:1;R2:1;R3:1;R4:1;R… TRUE      TRUE      NOTHI… 1      
#>  2 adenoma_1 APC       R1:1;R2:1;R3:1;R4:1;R… TRUE      TRUE      NOTHI… 1      
#>  3 adenoma_2 APC       R1:1;R2:1;R3:1;R4:1    TRUE      TRUE      NOTHI… 1      
#>  4 adenoma_2 KRAS      R1:0;R2:0;R3:0;R4:1    FALSE     TRUE      NOTHI… 2      
#>  5 adenoma_2 ARID2     R1:0;R2:0;R3:0;R4:1    FALSE     TRUE      NOTHI… 2      
#>  6 adenoma_2 TP53      R1:1;R2:1;R3:1;R4:0    FALSE     TRUE      NOTHI… 3      
#>  7 adenoma_3 PIK3CA    R1:0;R2:0;R3:0;R4:1;R… FALSE     TRUE      NOTHI… 1      
#>  8 adenoma_3 FBXW7     R1:1;R2:1;R3:1;R4:1;R… TRUE      TRUE      NOTHI… 2      
#>  9 adenoma_3 APC       R1:1;R2:1;R3:1;R4:1;R… TRUE      TRUE      NOTHI… 2      
#> 10 adenoma_3 AKAP9     R1:0;R2:0;R3:1;R4:0;R… FALSE     TRUE      NOTHI… 3      
#> # … with 53 more rows