Multi-region whole-genome sequencing data (median coverage ~100x) for the Set6 sample of the MSeq cohort, originally sequenced at lower resolution by Cross et al. (PMID 30177804). In this new version of the data (doi:10.1101/586560) we provide somatic single nucleotide variants, absolute clonal copy number segments and purity estimates.

Multi-region whole-genome sequencing data (median coverage ~100x) for the Set6 sample of the MSeq cohort, originally sequenced at lower resolution by Cross et al. (PMID 30177804). In this new version of the data (doi:10.1101/586560) we provide somatic single nucleotide variants, absolute clonal copy number segments and purity estimates.

data(MSEQ_CRC_ADENOCARCINOMA_SET6)

Format

A list that contains a tibble with mutation, copy number and purity data.

References

G. Caravagna, T. Heide, M. Williams, L. Zapata, D. Nichol, K. Chkhaidze, W. Cross, G.D. Cresswell, B. Werner, A. Acar, C.P. Barnes, G. Sanguinetti, T.A. Graham, A. Sottoriva. Model-based tumor subclonal reconstruction. biorxiv 2019, doi:10.1101/586560

See also

Examples

data(MSEQ_CRC_ADENOCARCINOMA_SET6)
MSEQ_CRC_ADENOCARCINOMA_SET6

#> $mutations
#> # A tibble: 28,840 x 35
#>    chr     from     to ref   alt   gene  region cosmic function. mutlocation
#>    <chr>  <dbl>  <dbl> <chr> <chr> <chr> <chr>  <lgl>  <chr>     <chr>      
#>  1 chr2  1.33e8 1.33e8 C     T     ANKR… inter… NA     NA        NA         
#>  2 chr2  1.33e8 1.33e8 C     T     ZNF8… inter… NA     NA        NA         
#>  3 chr2  1.33e8 1.33e8 C     T     ZNF8… inter… NA     NA        NA         
#>  4 chr2  1.33e8 1.33e8 C     T     ZNF8… inter… NA     NA        NA         
#>  5 chr2  1.33e8 1.33e8 A     T     GPR39 intro… NA     NA        NA         
#>  6 chr2  1.33e8 1.33e8 A     T     GPR39 intro… NA     NA        NA         
#>  7 chr2  1.33e8 1.33e8 G     T     GPR39 intro… NA     NA        NA         
#>  8 chr2  1.34e8 1.34e8 A     G     MIR7… ncRNA… NA     NA        NA         
#>  9 chr2  1.34e8 1.34e8 A     T     MIR7… ncRNA… NA     NA        NA         
#> 10 chr2  1.34e8 1.34e8 G     A     MIR7… ncRNA… NA     NA        NA         
#> # … with 28,830 more rows, and 25 more variables: patient <chr>,
#> #   Set6_42.VAF <dbl>, Set6_44.VAF <dbl>, Set6_45.VAF <dbl>, Set6_46.VAF <dbl>,
#> #   Set6_47.VAF <dbl>, Set6_48.VAF <dbl>, Set6_42.DP <dbl>, Set6_44.DP <dbl>,
#> #   Set6_45.DP <dbl>, Set6_46.DP <dbl>, Set6_47.DP <dbl>, Set6_48.DP <dbl>,
#> #   Set6_42.NV <dbl>, Set6_44.NV <dbl>, Set6_45.NV <dbl>, Set6_46.NV <dbl>,
#> #   Set6_47.NV <dbl>, Set6_48.NV <dbl>, Set6_42_N.VAF <dbl>,
#> #   Set6_44_N.VAF <dbl>, Set6_45_N.VAF <dbl>, Set6_46_N.VAF <dbl>,
#> #   Set6_47_N.VAF <dbl>, Set6_48_N.VAF <dbl>
#> 
#> $CNA
#> # A tibble: 85 x 15
#>    chr     from     to Set6_42.minor Set6_42.Major Set6_44.minor Set6_44.Major
#>    <chr>  <dbl>  <dbl>         <dbl>         <dbl>         <dbl>         <dbl>
#>  1 chr1  1.42e6 2.54e6             1             1             1             1
#>  2 chr1  3.43e6 4.71e7             0             1             0             1
#>  3 chr1  4.87e7 2.48e8             1             1             1             1
#>  4 chr1  2.48e8 2.48e8             0             2             0             2
#>  5 chr2  2.25e5 2.43e8             1             1             1             1
#>  6 chr3  2.91e6 1.97e8             1             1             1             1
#>  7 chr4  8.18e5 1.74e6             1             1             1             1
#>  8 chr4  1.75e6 1.06e7             1             1             1             1
#>  9 chr4  1.35e7 1.91e8             1             1             1             1
#> 10 chr4  1.91e8 1.91e8             1             2             1             2
#> # … with 75 more rows, and 8 more variables: Set6_45.minor <dbl>,
#> #   Set6_45.Major <dbl>, Set6_46.minor <dbl>, Set6_46.Major <dbl>,
#> #   Set6_47.minor <dbl>, Set6_47.Major <dbl>, Set6_48.minor <dbl>,
#> #   Set6_48.Major <dbl>
#> 
#> $samples
#> [1] "Set6_42" "Set6_44" "Set6_45" "Set6_46" "Set6_47" "Set6_48"
#> 
#> $purity
#> Set6_42 Set6_44 Set6_45 Set6_46 Set6_47 Set6_48 
#>    0.66    0.72    0.80    0.80    0.80    0.80 
#>