Multi-region whole-genome sequencing data (median coverage ~100x) for the Set7 sample of the MSeq cohort, originally sequenced at lower resolution by Cross et al. (PMID 30177804). In this new version of the data (doi:10.1101/586560) we provide somatic single nucleotide variants, absolute clonal copy number segments and purity estimates.

Multi-region whole-genome sequencing data (median coverage ~100x) for the Set7 sample of the MSeq cohort, originally sequenced at lower resolution by Cross et al. (PMID 30177804). In this new version of the data (doi:10.1101/586560) we provide somatic single nucleotide variants, absolute clonal copy number segments and purity estimates.

data(MSEQ_CRC_ADENOCARCINOMA_SET7)

Format

A list that contains a tibble with mutation, copy number and purity data.

References

G. Caravagna, T. Heide, M. Williams, L. Zapata, D. Nichol, K. Chkhaidze, W. Cross, G.D. Cresswell, B. Werner, A. Acar, C.P. Barnes, G. Sanguinetti, T.A. Graham, A. Sottoriva. Model-based tumor subclonal reconstruction. biorxiv 2019, doi:10.1101/586560

See also

Examples

data(MSEQ_CRC_ADENOCARCINOMA_SET7)
MSEQ_CRC_ADENOCARCINOMA_SET7

#> $mutations
#> # A tibble: 52,274 x 23
#>    chr     from     to ref   alt   gene  region cosmic function. mutlocation
#>    <chr>  <dbl>  <dbl> <chr> <chr> <chr> <chr>  <chr>  <chr>     <chr>      
#>  1 chr1  8.36e7 8.36e7 C     T     LINC… inter… NA     NA        NA         
#>  2 chr16 4.84e7 4.84e7 C     A     MIR5… ncRNA… NA     NA        NA         
#>  3 chr16 4.88e7 4.88e7 G     T     MIR5… inter… NA     NA        NA         
#>  4 chr16 4.89e7 4.89e7 T     G     MIR5… inter… NA     NA        NA         
#>  5 chr16 4.92e7 4.92e7 A     C     MIR5… inter… NA     NA        NA         
#>  6 chr16 4.92e7 4.92e7 T     C     MIR5… inter… NA     NA        NA         
#>  7 chr16 4.92e7 4.92e7 A     G     MIR5… inter… NA     NA        NA         
#>  8 chr16 4.92e7 4.92e7 G     A     MIR5… inter… NA     NA        NA         
#>  9 chr16 4.92e7 4.92e7 C     G     MIR5… inter… NA     NA        NA         
#> 10 chr16 4.92e7 4.92e7 A     T     MIR5… inter… NA     NA        NA         
#> # … with 52,264 more rows, and 13 more variables: patient <chr>,
#> #   Set7_55.VAF <dbl>, Set7_57.VAF <dbl>, Set7_59.VAF <dbl>, Set7_62.VAF <dbl>,
#> #   Set7_55.DP <dbl>, Set7_57.DP <dbl>, Set7_59.DP <dbl>, Set7_62.DP <dbl>,
#> #   Set7_55.NV <dbl>, Set7_57.NV <dbl>, Set7_59.NV <dbl>, Set7_62.NV <dbl>
#> 
#> $CNA
#> # A tibble: 60 x 11
#>    chr     from     to Set7_55.minor Set7_55.Major Set7_57.minor Set7_57.Major
#>    <chr>  <dbl>  <dbl>         <dbl>         <dbl>         <dbl>         <dbl>
#>  1 chr1  1.58e6 2.26e8             1             1             1             1
#>  2 chr1  2.27e8 2.27e8             1             1             0             2
#>  3 chr1  2.27e8 2.48e8             1             1             0             1
#>  4 chr2  2.80e5 2.42e8             1             1             1             1
#>  5 chr3  3.86e5 1.41e8             1             1             1             1
#>  6 chr3  1.41e8 1.81e8             1             2             1             2
#>  7 chr3  1.83e8 1.97e8             1             1             1             1
#>  8 chr4  7.20e5 6.32e6             1             1             1             1
#>  9 chr4  7.70e6 1.88e8             1             1             1             1
#> 10 chr5  1.71e6 6.51e7             1             1             1             1
#> # … with 50 more rows, and 4 more variables: Set7_59.minor <dbl>,
#> #   Set7_59.Major <dbl>, Set7_62.minor <dbl>, Set7_62.Major <dbl>
#> 
#> $samples
#> [1] "Set7_55" "Set7_57" "Set7_59" "Set7_62"
#> 
#> $purity
#> Set7_55 Set7_57 Set7_59 Set7_62 
#>    0.88    0.88    0.88    0.80 
#>