Mutations and matched copy number from TRACERx lung (PMID: 28445112). n = 99 patients, multi-region whole-exome sequencing. Calls generated from Jamal-Hanjani et al. (PMID: 28445112).

Mutations and matched copy number from TRACERx lung (PMID: 28445112). n = 99 patients, multi-region whole-exome sequencing. Calls generated from Jamal-Hanjani et al. (PMID: 28445112).

data(TRACERx_NEJM_2017)

Format

Data for the TRACERx cohort. Somatic mutations (~65K) from >300 tumour biopsies, whole-exome sequenced at high-depth (~400x). For every mutation the Cancer Fraction is reported - VAF adjusted for tumour content and copy number data. All the data (calls, definition of driver events, etc.) is as released in the original paper. The data is as used in the paper Caravagna et al (PMID: 30171232) to show repeated evolution in lung cancers with REVOLVER.

References

Jamal-Hanjani, M., Wilson, G.A., McGranahan, N., Birkbak, N.J., Watkins, T.B., Veeriah, S., Shafi, S., Johnson, D.H., Mitter, R., Rosenthal, R., et al., 2017. Tracking the evolution of non–small-cell lung cancer. New England Journal of Medicine, 376(22), pp.2109-2121. PMID 28445112.

Examples

data(TRACERx_NEJM_2017)
TRACERx_NEJM_2017

#> # A tibble: 65,421 x 7
#>    Misc             patientID variantID cluster is.driver is.clonal CCF         
#>    <chr>            <chr>     <chr>     <chr>   <lgl>     <lgl>     <chr>       
#>  1 CRUK0001:7:4797… CRUK0001  PKD1L1    1       FALSE     FALSE     R1:0.86;R2:…
#>  2 CRUK0001:21:477… CRUK0001  PCNT      1       FALSE     FALSE     R1:0.86;R2:…
#>  3 CRUK0001:22:373… CRUK0001  CSF2RB    1       FALSE     FALSE     R1:0.86;R2:…
#>  4 CRUK0001:10:102… CRUK0001  KAZALD1   1       FALSE     FALSE     R1:0.86;R2:…
#>  5 CRUK0001:7:8197… CRUK0001  CACNA2D1  1       FALSE     FALSE     R1:0.86;R2:…
#>  6 CRUK0001:1:2018… CRUK0001  IPO9      1       FALSE     FALSE     R1:0.86;R2:…
#>  7 CRUK0001:3:1674… CRUK0001  PDCD10    1       FALSE     FALSE     R1:0.86;R2:…
#>  8 CRUK0001:7:1346… CRUK0001  CALD1     1       FALSE     FALSE     R1:0.86;R2:…
#>  9 CRUK0001:17:769… CRUK0001  CANT1     1       FALSE     FALSE     R1:0.86;R2:…
#> 10 CRUK0001:16:166… CRUK0001  IFT140    1       FALSE     FALSE     R1:0.86;R2:…
#> # … with 65,411 more rows