Mutations and copy number from breast cancer patients (PMID: 26099045). n = 50 patients, multi-region whole-exome sequencing. Calls generated from Yates et al. (PMID: 26099045).

Mutations and copy number from breast cancer patients (PMID: 26099045). n = 50 patients, multi-region whole-exome sequencing. Calls generated from Yates et al. (PMID: 26099045).

data(YATES_BREAST_NATMED_2015)

Format

Data for the breast cancer cohort by Yates et al. The data has been extracted from the Supplementary tables released in the original Nature Medicine paper (PMID: 26099045). Mutations and copy number are reported as present or absent in each one of the sequenced biopsies. All the data (calls, definition of driver events, etc.) is as released in the original paper. The data is as used in the paper Caravagna et al (PMID: 30171232) to show repeated evolution in breast cancers with REVOLVER.

References

Yates, L.R., Gerstung, M., Knappskog, S., Desmedt, C., Gundem, G., Van Loo, P., Aas, T., Alexandrov, L.B., Larsimont, D., Davies, H. et al., 2015. Subclonal diversification of primary breast cancer revealed by multiregion sequencing. Nature Medicine, 21(7), p.751. PMID: 26099045

Examples

data(YATES_BREAST_NATMED_2015)
YATES_BREAST_NATMED_2015

#> # A tibble: 403 x 7
#>    Misc   patientID variantID  cluster is.driver is.clonal CCF                  
#>    <chr>  <chr>     <chr>      <chr>   <lgl>     <lgl>     <chr>                
#>  1 UNKNO… PD9193    JAK3       1       FALSE     TRUE      PD9193d:1;PD9193e:1;…
#>  2 UNKNO… PD9193    BLM        2       FALSE     FALSE     PD9193d:1;PD9193e:1;…
#>  3 UNKNO… PD9193    SMARCA4    3       TRUE      FALSE     PD9193d:1;PD9193e:0;…
#>  4 UNKNO… PD9193    GATA3      1       TRUE      TRUE      PD9193d:1;PD9193e:1;…
#>  5 UNKNO… PD9193    +8q        4       TRUE      FALSE     PD9193d:1;PD9193e:0;…
#>  6 UNKNO… PD9193    -8p        4       TRUE      FALSE     PD9193d:1;PD9193e:0;…
#>  7 UNKNO… PD9694    SF3B1      1       FALSE     TRUE      PD9694e:1;PD9694f:1;…
#>  8 UNKNO… PD9694    CREBBP     1       TRUE      TRUE      PD9694e:1;PD9694f:1;…
#>  9 UNKNO… PD9694    SETD2      2       TRUE      FALSE     PD9694e:1;PD9694f:0;…
#> 10 UNKNO… PD9694    amp(8p11.… 1       TRUE      TRUE      PD9694e:1;PD9694f:1;…
#> # … with 393 more rows