Print the list of available datasets.

For each dataset a code and brief descriptions are reported.

available_data()

Value

Nothing.

Examples

available_data()

#> 

#> ── TRACERx_NEJM_2017 ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

#> ● Mutations from the TRACERx (http://tracerx.co.uk/) lung cohort, as it was released in Jamal-Hanjani et al. N Engl J Med 2017; 376:2109-2121 (DOI: 10.1056/NEJMoa1616288).
#> ● Multi-region high-resolution whole exome data of n = 99 cancer patients.
#> ● Primary non-small cell lung adenocarcinomas and squamous carcinomas.
#> ● Somatic mutations data with Cancer Cell Fractions (purity and copy number adjusted variant allele frequency).
#> 

#> ── TRACERx_NEJM_2017_REVOLVER ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

#> ● REVOLVER (https://caravagn.github.io/revolver) analysis of the TRACERx (http://tracerx.co.uk/) lung cohort available in thi package.
#> ● The original input data is available in this package with codename: TRACERx_NEJM_2017.
#> 

#> ── YATES_BREAST_NATMED_2015 ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

#> ● Mutations from a breast cancer cohort, as it was released in Yates et al. Nat Med 21, 751–759 (2015) (PMID: 26099045).
#> ● Multi-region high-resolution whole genome/ targeted data of n = 50 cancer patients.
#> ● Primary breast cancers.
#> ● Binary (present/ absent) somatic mutations data for putative driver events.
#> 

#> ── CROSS_CRC_ADENOCARCINOMA_NATECOEVO_2018 ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

#> ● Mutations from a colorectal cancer cohort, as it was released in Cross et al. Nat Ecol Evol 2(10):1661-1672, (2018) (doi:10.1038/s41559-018-0642-z).
#> ● Multi-region high-resolution whole exome data of n = 19 cancer patients.
#> ● Primary adenomas and colorectal carcinomas.
#> ● Binary (present/ absent) somatic mutations data for putative driver events.
#> 

#> ── MSEQ_CRC_ADENOCARCINOMA_SET6 ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

#> ● Mutations, Copy Number Alteration and tumour purity from a colorectal cancer patient, first used in Caravagna et al. biorXiv, (2019) (doi:10.1101/586560).
#> ● Multi-region high-resolution whole genome (~100x) data of cancer patient 'Set6'.
#> ● Multiple samples of a colorectal adenocarcinoma, re-sequencing of a sample first analysed in Cross et al. Nat Ecol Evol 2(10):1661-1672, (2018) (doi:10.1038/s41559-018-0642-z).
#> ● List of Somatic mutations data, clonal absolute copy number data (segments format), an tumour purity.
#> 

#> ── MSEQ_CRC_ADENOCARCINOMA_SET7 ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

#> ● Mutations, Copy Number Alteration and tumour purity from a colorectal cancer patient, first used in Caravagna et al. biorXiv, (2019) (doi:10.1101/586560).
#> ● Multi-region high-resolution whole genome (~100x) data of cancer patient 'Set7'.
#> ● Multiple samples of a colorectal adenocarcinoma, re-sequencing of a sample first analysed in Cross et al. Nat Ecol Evol 2(10):1661-1672, (2018) (doi:10.1038/s41559-018-0642-z).
#> ● List of somatic mutations data, clonal absolute copy number data (segments format), an tumour purity.